pyproject.toml
[tool.poetry]
name = "pyfsm"
version = "0.1.0"
description = ""
authors = ["nyvemm <joaovictorsawada@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^1.4.46"
alembic = "^1.9.1"
python-dotenv = "^0.21.0"
flask-sqlalchemy = "^3.0.2"
flask-restx = "^1.0.5"
flask-caching = "^2.0.2"
unittest2 = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
